// src/generated/model/utbetalingerFrontend.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface UtbetalingerFrontend {
    bekreftelse?: boolean;
    utbytte: boolean;
    salg: boolean;
    forsikring: boolean;
    annet: boolean;
    beskrivelseAvAnnet?: string;
    utbetalingerFraNavFeilet?: boolean;
}